requirements.txt
Django==5.1.6
django-ninja==1.3.0
gunicorn==21.2.0
mysqlclient==2.2.7
django-cors-headers==4.3.1
